Simplify EventSheetService interface and mockup
Reduce the number of features in the interface to just adding and
removing events, and editing event metadata. This corresponds to the
data stored in the event metadata sheet only, leaving event signups
to be defined and implemented at a later date.

Simplify the mockup by merging it into the events sheet service with
dummy methods that do not preserve state, rather than the more complex
and fully featured mock service that was introduced in the previous
commit.