Add ability to add events to sheet
To do this: add tracking of largest key in use, and get this from the
sheet on service construction. Also, add a dictionary to map between
event keys and rows; this will come in useful when editing and
deleting events, and make sure the event adding method keeps these
up to date.

For context, when events are deleted either the will be blank rows in
the sheet, or the event key will not correspond to the row number of
the event. I chose the second option, so we need to keep track of this
mapping.