Sheets service: remove events (#16)
Extract metadata sheet name into a constant.

Add initialisation of metadata sheet id. For queries so far we have
been able to use its title, but to delete it we need to know its
sheetId. This does not change, and can be established whilst the
service is being constructed.

Implement remove events method.