Add interface for batching event responses
Add a thread-safe buffer for storing reactions as they come in, and
process them up to every 10s.

Add an interface to the EventsSheetsService to process these batches
of requests. Add a model to store all information needed in a reaction:
a message (ie event) ID, a user ID and a reaction emoji (response
type).